change data pattern to context provider
