+++ src/lib/context.ts
@@ -1,0 +1,12 @@
+import { createContext } from "svelte"
+import type { BuildingData, ClassMapValue, CollegeData, DivisionData, RoomData } from "./types";
+
+type AppData = {
+  rooms: RoomData[];
+  buildings: BuildingData[];
+  colleges: CollegeData[];
+  divisions: DivisionData[];
+  classesMap: Map<string, ClassMapValue[]>;
+}
+
+export const [getAppData, setAppData] = createContext<AppData>();
